Move mh_log to public header location
Will obviously need to undo this later
