+++ src/mono/mono/metadata/CMakeLists.txt
@@ -101,7 +101,6 @@ set(metadata_common_sources
     method-builder-internals.h
     method-builder.c
     mh_log.c
-    mh_log.h
     mono-basic-block.c
     mono-basic-block.h
     mono-config.c
