+++ src/mono/mono/metadata/mh_log.c
@@ -1,4 +1,4 @@
-#include "mh_log.h"
+#include <mono/metadata/mh_log.h>
 #include <mono/metadata/metadata-internals.h>
