+++ src/native/public/mono/metadata/mh_log.h
